modularization: extract text generator
